requirements.txt
datasets
tokenizers
openai
transformers
python-dotenv
scikit-learn
psycopg2
langchain
uvicorn
fastapi
langchain_community
langchain_openai
langchain_postgres
langchain_azure_ai

az-cli
azure.identity
azure-search-documents
azure-ai-inference
azure-cosmos>4.4